Refactor workflow canvas credentials
