+++ apps/canvas/src/features/workflow/pages/workflow-canvas/constants.ts
@@ -1,0 +1,5 @@
+export const HISTORY_LIMIT = 50;
+export const WORKFLOW_CLIPBOARD_HEADER = "ORCHEO_WORKFLOW_CLIPBOARD_V1:";
+export const PASTE_BASE_OFFSET = 40;
+export const PASTE_OFFSET_INCREMENT = 24;
+export const PASTE_OFFSET_MAX_STEPS = 5;
