+++ apps/canvas/src/features/workflow/pages/workflow-canvas/validators.ts
@@ -1,0 +1,97 @@
+import type {
+  NodeStatus,
+  WorkflowExecutionStatus,
+} from "./types";
+
+export const isRecord = (value: unknown): value is Record<string, unknown> => {
+  return typeof value === "object" && value !== null;
+};
+
+export const determineNodeType = (nodeId?: string) => {
+  if (nodeId?.includes("chat-trigger")) {
+    return "chatTrigger" as const;
+  }
+  if (nodeId === "sticky-note") {
+    return "stickyNote" as const;
+  }
+  if (nodeId === "start-node" || nodeId === "end-node") {
+    return "startEnd" as const;
+  }
+  return "default" as const;
+};
+
+export const validateWorkflowData = (data: unknown) => {
+  if (!isRecord(data)) {
+    throw new Error("Invalid workflow file structure.");
+  }
+
+  const { nodes, edges } = data;
+
+  if (!Array.isArray(nodes)) {
+    throw new Error("Invalid nodes array in workflow file.");
+  }
+
+  nodes.forEach((node, index) => {
+    if (!isRecord(node)) {
+      throw new Error(`Invalid node at index ${index}.`);
+    }
+    if (!isRecord(node.position)) {
+      throw new Error(`Node ${node.id ?? index} is missing position data.`);
+    }
+    const { x, y } = node.position as Record<string, unknown>;
+    if (typeof x !== "number" || typeof y !== "number") {
+      throw new Error(`Node ${node.id ?? index} has invalid coordinates.`);
+    }
+  });
+
+  if (!Array.isArray(edges)) {
+    throw new Error("Invalid edges array in workflow file.");
+  }
+
+  edges.forEach((edge, index) => {
+    if (!isRecord(edge)) {
+      throw new Error(`Invalid edge at index ${index}.`);
+    }
+    if (typeof edge.source !== "string" || typeof edge.target !== "string") {
+      throw new Error(`Edge ${edge.id ?? index} has invalid connections.`);
+    }
+  });
+};
+
+export const nodeStatusFromValue = (value?: string): NodeStatus => {
+  const normalised = value?.toLowerCase();
+  switch (normalised) {
+    case "running":
+      return "running";
+    case "error":
+    case "failed":
+      return "error";
+    case "warning":
+    case "cancelled":
+    case "partial":
+      return "warning";
+    default:
+      return "success";
+  }
+};
+
+export const executionStatusFromValue = (
+  value?: string,
+): WorkflowExecutionStatus | null => {
+  const normalised = value?.toLowerCase();
+  switch (normalised) {
+    case "running":
+      return "running";
+    case "completed":
+    case "success":
+      return "success";
+    case "error":
+    case "failed":
+      return "failed";
+    case "cancelled":
+    case "partial":
+      return "partial";
+    default:
+      return null;
+  }
+};
